+++ CHANGELOG.md
@@ -4,15 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Fixed
+- Now melting curves show the proper (inverted) dissociation curve.
+
 ## [1.4.3]
 ### Added
 - Melting curve plotting script (from fish-conditions repo).
 
 ### Changed
 - Moved functions to separate library
-
-### Fixed
-- Now melting curves show the proper (inverted) dissociation curve.
 
 ## [1.4.2]
 ### Changed
@@ -47,6 +48,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [1.0.0]
 
+* [unreleased] https://github.com/ggirelli/oligo-melting
 * [1.4.3] https://github.com/ggirelli/oligo-melting/releases/tag/v1.4.3
 * [1.4.2] https://github.com/ggirelli/oligo-melting/releases/tag/v1.4.2
 * [1.4.0] https://github.com/ggirelli/oligo-melting